Marshaling.

+++ scanner/src/protocol.rs
@@ -179,7 +179,7 @@ impl Type {
         match *self {
             Type::Int => "i32",
             Type::Uint => "u32",
-            Type::Fixed => "f32",
+            Type::Fixed => "f64",
             Type::Array => "wl_array",
             Type::Fd => "::std::os::unix::io::RawFd",
             Type::String => "String",
